Update .jar File

+++ PostService/src/test/java/org/thewhitemage13/controller/PostControllerTest.java
@@ -11,7 +11,7 @@ import org.springframework.http.ResponseEntity;
 import org.thewhitemage13.dto.CreatePostDTO;
 import org.thewhitemage13.dto.OpenPostDTO;
 import org.thewhitemage13.dto.UpdatePostDTO;
-import org.thewhitemage13.exception.PostNotFoundException;
+import org.thewhitemage13.exceptions.PostNotFoundException;
 import org.thewhitemage13.service.PostServiceImpl;
 
 import java.util.List;
@@ -119,7 +119,7 @@ class PostControllerTest {
         Mockito.doReturn(postCount).when(postServiceImpl).getCountPostByUserId(userId);
 
         // when
-        ResponseEntity<Long> response = postController.getPostCountByUserId(userId);
+        ResponseEntity<Integer> response = postController.getPostCountByUserId(userId);
 
         // then
         assertNotNull(response);
